First challenge of the tutorial
Implement guard waypoint walking on your own.

Used the coRoutine video as a reference and wrote up a quick one to
walk the given waypoints. Seems pretty simple, but there's probably a
better way to have the guard walk and search forever than using a "while
true" in here. It reminds me of state machines, in that we could
probably program behavior for the guard based on some state like
searching or hunting kinda thing. That'd probably be really fun, make
the guards chase the player if they spot them or something like that.

Anyway, time to watch more of the tutorial and then see how Sebastian
implements the solution.