Make Guard rotate towards waypoint as they patrol
Started the challenge on my own, knew I needed atan2 and some eulerAngle
assignment, but had to watch Sebastian's vid to actually find out how to
determine the angle to look at first. Of course it's jut a target -
transform.position normalized vector... I guess? It's been a while
since I did linear algebra, but I guess it makes sense that taking the
difference between two vectors would result in a way to move towards one
of them.

One thing I was right about before watching his implementation though
was that we'd need to start the coroutine to rotate when we got to the
waypoint, but I didn't realize we could have more than one yield return
in a single CoRoutine. How weird.